Cambios en consulta de asistencias, agregado asignaciones y descuentos.

+++ BugTrackingSystem/BugTrackingSystem/Entidades/SueldoAsignacion.cs
@@ -7,7 +7,7 @@ namespace BugTrackingSystem.Entidades
         public Usuario Usuario { get; set; }
         public DateTime Fecha { get; set; }
         public Asignacion Asignacion { get; set; }
-        public float Monto { get; set; }
+        public decimal Monto { get; set; }
         public int Cantidad { get; set; }
         public bool Borrado { get; set; }
     }
